Replace kameo with our own code
This tries to not change to much of the code using kameo as possible,
but makes concessions where it would be much easier to change the API.

This keeps the same API for the most part but has mostly small changes.
The biggest change is the removal of the Reply trait, and the return
value of Message::handle. Instead replies are not de-structured at all
they all must be sent via Context::reply_sender / Context::reply