Send actor reply to ask is now an async function
This will let me add a feature where we send a message to an actor as a
reply.

This includes a functionality change. If an actor panics when handling
an ask message, the user waiting for the reply will now no longer get
NoReply and instead get ActorStopped, which seems good actually.