Avoid unbounded host buffer allocation during execution (#3545)
In various ecall and syscalls, the guest passes a pointer and length to
a region in memory that should be read by the host. Currently on `main`,
many of these allocate a `Vec<u8>` to copy the region pointed to by the
guest. This opens up an issue for provers that need to execute arbitrary
guest programs (e.g. on Boundless) in that the guest can supply a very
large (up to 4GB) size to read, and the host will allocate a large
buffer. This can be used to cause an OOM condition in the executor.

When running unknown guest programs, limits should be placed on the
amount of memory the process can use (e.g. via `cgroups`) and in
practice this is done. This mitigates the most significant possible
effects, but this kind of issue can still cause friction.

This PR addresses this issue by providing the ability to read from guest
memory without copying, through the `std::io::Read` trait. This is then
used in all relevant callsites to avoid allocating a large buffer unless
required (e.g. the journal needs to be read into a buffer that will be
returned with the session). At those callsites, limits are added on how
much can be read.

This PR is built on top of https://github.com/risc0/risc0/pull/3512