m3 executor splitting (#3552)
This PR implements a new segment splitting logic for m3.

It counts the number of "row points" that is portions of a row that each
block takes and uses this to keep track of when to split.

This creates a one-to-one relationship between the executor segments and
the m3 preflight segments again. To make this work I had to reverse some
multi-segment code in m3. Some multi-segment code here remains still
like the preflight iterator, we can remove this in a future PR.

The one-to-one relationship fixes the actor-system to work with m3 now.

po2 > 20 doesn't seem to work for m3 yet for unknown reasons.