Add epoch counter to optimize reset of PageTable (#3563)
In my profiling of execution, I have seen that the `PageTable::reset`
function takes up a significant portion of overall execution time, often
around 5-10%. The reset is responsible for producing a zeroed
`PageTable` vector, which is a contiguous chunk of memory with one `u32`
per page, indexing into the cache of loaded `Page` objects.

This PR is a proposed optimization. It essentially eliminates the reset
cost by only reallocating the vector every 1024 segments, incrementing
an "epoch" number instead to invalid past entries. I'm opening this PR
to share the idea.