Refactor executor APIs to allow for deferred hashing memory images (#3514)
In the current API, inclusion of `Segment` and `ReceiptClaim` force
hashing of memory images when it is not required. In order to allow for
deferred hashing of memory images (e.g. to preflight) and for execution
entirely without hashing (e.g. in dev mode), this PR refactors the
executor APIs.

In addition to moving the construction of `Segment` and `ReceiptClaim`
types, it makes the `risc0-zkvm` and `risc0-circuit-rv32im` executor
implementations more compatible with iterator-like execution.

This PR includes changes the `MemoryImage` and `Segment` to allow the
executor to avoid fully hashing the memory image on each segment split,
instead only computing the hashes required to provide the partial memory
image to the preflight (the Merkle tree uncle nodes).

Additionally, it moves the code that handles creating a dedicated
hashing thread into `risc0_zkvm` and remove hashing entirely from the
execution-only code paths.

This is a follow-up to https://github.com/risc0/risc0/pull/3512