Enable m3 by default (#3585)
This fixes lingering issues with m3 and enables the feature-flag by
default

In order to make this work, I had to do two regrettable things:
- disable creation of po2=13 final segments in executor, it reveals
issues with block tracker that need a deeper fix.
- disable the POVW test since m3 segments don't have support for it yet

I think it is worthwhile to move forward with this PR ahead of fixing
those issues to swap our CI coverage over to m3 so we can avoid breaking
m3 with further commits.