Make the claim Merge trait pub and add prune methods (#3589)
In some work on Boundless a couple of months ago, and now on the
`risc0-r0vm` crate I came across cases where I wanted to be able to
merge two `MaybePruned` objects, which exists as a private trait. I
worked around the issue in both cases, but I think it would be nice to
have this trait public as I think its quite useful and is also fairly
small.

This PR also adds a `MaybePruned::prune` method. This can be used in
cases where you want less data. It is useful if you want to send a
`ReceiptClaim` over the wire, and want to make sure you don't
accidentally send an arbitrarily large journal it may contain.