Remove rv32im-m3 feature flag (#3598)
This removes the rv32im-m3 feature flag.

This flag is now on by default, so what this does is remove the ability
to turn off m3 (and thus turn on v2).

This causes some code to be dead but I will remove it and collapse the
rv32im-m3 crates in a future PR.